requirements.txt
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
distlib==0.3.4
dnspython==2.2.0
email-validator==1.1.3
filelock==3.4.2
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
platformdirs==2.4.1
pycparser==2.21
pymongo==4.0.2
six==1.16.0
SQLAlchemy==1.4.32
urllib3==1.26.8
virtualenv==20.13.3
Werkzeug==2.0.3
WTForms==3.0.1
